Move response stuff into folder
